Add parameter to create own download links
Using the new flag -prefix=..., you can now change the links in the file
plugin-center.json to match a static server url where you can deploy the
plugin file structure created by this program. With this it is possible,
to just host all the files statically eg. in an nginx an use the url
for the plugin-center.json file as the plugin center url in SCM-Manager.